replace auto increment with timestamp global sequence

+++ run/sql.mysql/tableCreates.sql
@@ -56,8 +56,8 @@ create table bmsql_customer (
 -- create sequence bmsql_hist_id_seq;
 
 create table bmsql_history (
-  hist_id integer not null auto_increment primary key,
-  h_c_id   bigint,
+  hist_id bigint not null auto_increment primary key,
+  h_c_id   integer,
   h_c_d_id integer,
   h_c_w_id integer,
   h_d_id   integer,
